Drop the Queries module
Hat tip to Jamie and Lasitha for the direction <3
