--- app/queries/items_by_storage_collection.rb
@@ -1,31 +1,0 @@
-module Query
-  class ItemsByStorageCollection
-    attr_reader :organization
-    attr_reader :filter_params
-
-    def initialize(organization:, filter_params:)
-      @organization = organization
-      @filter_params = filter_params
-    end
-
-    # rubocop:disable Naming/MemoizedInstanceVariableName
-    def call
-      @items ||=  organization
-                  .items
-                  .joins(' LEFT OUTER JOIN "inventory_items" ON "inventory_items"."item_id" = "items"."id"')
-                  .joins(' LEFT OUTER JOIN "storage_locations" ON "storage_locations"."id" = "inventory_items"."storage_location_id"')
-                  .select('
-                          items.id,
-                          items.name,
-                          items.category,
-                          items.barcode_count,
-                          storage_locations.name as storage_name,
-                          storage_locations.id as storage_id,
-                          sum(inventory_items.quantity) as quantity
-                        ')
-                  .group("storage_locations.name, storage_locations.id, items.id, items.name")
-                  .order(name: :asc).filter(filter_params)
-    end
-    # rubocop:enable Naming/MemoizedInstanceVariableName
-  end
-end
